+++ cyberdrop_dl/prompts.py
@@ -115,3 +115,8 @@ def ask_path(
             console.error(repr(e))
         else:
             return path.resolve()
+
+
+def ask_should_create_config(file: Path) -> bool:
+    console.warning("A default config file does not exists")
+    return ask_confirmation(f"Do you want to create it at '{file}'?")